Create db_autopwn.rb
db_autopwn command was removed from newer metasploit framework versions. 
I do consider this a valid tool, and found some adaptation of it to run as msf plugin on the web. The adapted code was just not working right on current msf. This new source code is corrected and tested to msf 4.5.0.